Change method for process termination, still working on fixing it. Works fine for non-terminating functions

+++ linked_list.h
@@ -11,7 +11,7 @@
 #define PROCESS_RAM 256 // Each process will have 512 words of ram
 //typedef struct PCB PCB;
 typedef struct PCB {
-	uint16_t id; // The process id
+	uint8_t id; // The process id. This should map directly to the index of the array it's stored in
 	void (*function) (void); // The function itself
 	uint16_t *stackPointer; // Where the process's stack pointer points to
 	uint16_t ram[PROCESS_RAM]; // The processes' ram
